NYU QA Code Repository
This is one of three resources used for the QA effort.  This contains
only code and related documentation.  Code stored in this repository
may depend on data which is available upon request from Tom Martin and
stored in Google Docs.  Deliverables which utilize this data will be
published in a yet-unknown public location.